+++ src/main/java/kitchenpos/domain/OrderLineItem.java
@@ -5,6 +5,7 @@ import javax.persistence.Entity;
 import javax.persistence.ForeignKey;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
+import kitchenpos.menu.domain.Menu;
 
 @Entity
 public class OrderLineItem extends BaseEntity{
