+++ src/main/java/kitchenpos/menu/application/dto/MenuCreateRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.application.dto;
+package kitchenpos.menu.application.dto;
 
 import java.util.List;
