+++ src/main/java/kitchenpos/menu/application/dto/MenuGroupCreateRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.application.dto;
+package kitchenpos.menu.application.dto;
 
 public class MenuGroupCreateRequest {
