+++ src/main/java/kitchenpos/menu/domain/MenuGroup.java
@@ -1,7 +1,8 @@
-package kitchenpos.domain;
+package kitchenpos.menu.domain;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
+import kitchenpos.domain.BaseEntity;
 
 @Entity
 public class MenuGroup extends BaseEntity {
