+++ src/test/java/kitchenpos/application/TableGroupServiceTest.java
@@ -9,13 +9,12 @@ import kitchenpos.application.dto.OrderCreateRequest.OrderLineRequest;
 import kitchenpos.application.dto.TableGroupCreateRequest;
 import kitchenpos.application.dto.TableGroupCreateRequest.OrderTableRequest;
 import kitchenpos.application.dto.TableGroupResponse;
-import kitchenpos.dao.MenuRepository;
+import kitchenpos.menu.domain.MenuRepository;
 import kitchenpos.dao.OrderLineItemRepository;
 import kitchenpos.dao.OrderRepository;
 import kitchenpos.dao.OrderTableRepository;
 import kitchenpos.dao.TableGroupRepository;
 import kitchenpos.domain.OrderTable;
-import kitchenpos.domain.TableGroup;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
